backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
redis
apscheduler
pandas
numpy
requests
psycopg2-binary
psycopg2-binary
python-jose[cryptography]
passlib
bcrypt==3.2.2
python-multipart
yfinance
httpx
kagglehub
newsapi-python
tradingeconomics
openai

